feat: add deleteCard permissions

+++ controllers/cards.js
@@ -3,6 +3,7 @@ const ValidationError = require('../errors/validation-error');
 const ServerError = require('../errors/server-error');
 const NotFoundError = require('../errors/not-found-error');
 const CastError = require('../errors/cast-error');
+const ForbiddenError = require('../errors/forbidden-error');
 
 const getCards = (req, res, next) => {
   Card.find({})
@@ -31,13 +32,18 @@ const createCard = (req, res, next) => {
 const deleteCard = (req, res, next) => {
   const { cardId } = req.params;
 
-  Card.findByIdAndDelete(cardId)
-    .orFail(new Error('NotValidId'))
+  Card.findById(cardId)
+    .then((card) => {
+      if (card.owner.toString() !== req.user._id.toString()) {
+        throw new ForbiddenError('You have no permission to delete other users cards');
+      }
+
+      Card.findByIdAndDelete(cardId)
+        .catch(next);
+    })
     .then((card) => res.status(200).send(card))
     .catch((err) => {
-      if (err.message === 'NotValidId') {
-        throw new NotFoundError('Card not found');
-      } else if (err.name === 'CastError') {
+      if (err.name === 'CastError') {
         throw new CastError('Wrong card Id');
       } else {
         throw new ServerError(`Server error: ${err}`);
